Ignore environment variables which may break this build.
Gecko sets various environment variables so that rust crates can build
C dependencies using cc correctly. However, this crate is primarily
intended to be used as a build-dependency, and some of those variables
are intended for target builds, and cause issues as a result. So unset
those variables before building glsl-optimizer.