Fix build on OpenBSD
Like on FreeBSD/Dragonfly, OpenBSD requires pthread_np.h to be included for
pthread_set_name_np(). Mesa's u_thread.h includes this file when
PTHREAD_SETAFFINITY_IN_NP_HEADER is defined. While this will be defined in
Mesa's build system for FreeBSD due to the presence of pthread_setaffinity_np(),
it is not clear to me how pthread_np.h gets included on OpenBSD.

However, defining this for OpenBSD appears to fix the build, so let's just do
it.

Fixes #7